+++ crates/ruma-events/src/room/aliases.rs
@@ -71,4 +71,5 @@ impl RedactedStateEventContent for RedactedRoomAliasesEventContent {
 
 impl StaticEventContent for RedactedRoomAliasesEventContent {
     const TYPE: &'static str = RoomAliasesEventContent::TYPE;
+    type IsPrefix = <RoomAliasesEventContent as StaticEventContent>::IsPrefix;
 }